Move derivation show to separate file
